Create desc
Welcome to NumPy! NumPy (Numerical Python) is an open source Python library that’s used in almost every field of science and engineering. It’s the universal standard for working with numerical data in Python, and it’s at the core of the scientific Python and Py Data ecosystems. NumPy users include everyone from beginning coders to experienced researchers doing state-of-the-art scientific and industrial research and development. The NumPy API is used extensively in Pandas, SciPy, Matplotlib, scikit-learn, scikit-image and most other data science and scientific Python packages. It contains 1D,2D and multi-dimensional arrays and matrix data structures. It is very efficient and highly powerful which supplies an enormous library support and powerful data structures that guarantee efficient calculations and high-level mathematical functions.